move regi stration to the separate file. add some routes
