+++ front-end/src/App.jsx
@@ -1,101 +1,6 @@
-import { useState } from 'react';
+import AppRoutes from './AppRoutes.jsx';
 
 function App() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [password_confirmation, setPassword_confirmation] = useState('');
-  const [responseMessage, setResponseMessage] = useState(null);
-  const [isOrganization, setIsOrganization] = useState(false);
-
-  const handleSubmit = async (event) => {
-    event.preventDefault();
-
-    try {
-      const url_root = import.meta.env.VITE_REACT_URL;
-      const response = await fetch(`${url_root}auth`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          auth: {
-            email: email,
-            password: password,
-            password_confirmation: password_confirmation,
-            isOrganization: isOrganization,
-          },
-        }),
-      });
-
-      if (!response.ok) {
-        throw new Error(`Error: ${response.status}`);
-      }
-
-      const data = await response.json();
-      setResponseMessage('Registration successful!');
-      console.log(data);
-    } catch (error) {
-      console.error('Error:', error);
-      setResponseMessage('Registration failed');
-    }
-  };
-
-  return (
-    <div>
-      <h1 className="text-4xl text-blue-500 bg-red-500 p-4">Tailwind</h1>
-      <h1>Care Connect</h1>
-      {responseMessage && <p>{responseMessage}</p>}
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label>Email:</label>
-          <input type="email" value={email} onChange={(e) => setEmail(e.target.value)} required />
-        </div>
-        <div>
-          <label>Password:</label>
-          <input type="password" value={password} onChange={(e) => setPassword(e.target.value)} required />
-        </div>
-        <div>
-          <label>Password:</label>
-          <input
-            type="password"
-            value={password_confirmation}
-            onChange={(e) => setPassword_confirmation(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <div>
-            <label>
-              <input
-                type="radio"
-                name="organizationType"
-                value="true"
-                checked={isOrganization === 'true'}
-                onChange={(e) => setIsOrganization(e.target.value)}
-                required
-              />
-              Organization
-            </label>
-          </div>
-          <div>
-            <label>
-              <input
-                type="radio"
-                name="organizationType"
-                value="false"
-                checked={isOrganization === 'false'}
-                onChange={(e) => setIsOrganization(e.target.value)}
-                required
-              />
-              Volunteer
-            </label>
-          </div>
-        </div>
-
-        <button type="submit">Register</button>
-      </form>
-    </div>
-  );
+  return <AppRoutes />;
 }
-
 export default App;
