+++ front-end/src/components/pages/HomePage.jsx
@@ -1,0 +1,5 @@
+function HomePage() {
+  return <a href="/register">Register</a>;
+}
+
+export default HomePage;
